feat(hwinfo): aggregate hardware information in output
