+++ venus-worker/hwinfo/Cargo.toml
@@ -5,6 +5,16 @@ edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
+[workspace]
+members = [
+    "hwinfo",
+]
+
+[features]
+default = ["opencl"]
+opencl = ["rust-gpu-tools/opencl"]
+cuda = ["rust-gpu-tools/cuda"]
+
 [dependencies]
 anyhow = "1"
 hwloc2 = "2.2.0"
@@ -12,5 +22,4 @@ sysinfo = "0.23"
 byte-unit = { version = "4.0", default-features = false }
 strum = { version = "0.24", features = ["derive"] }
 term-table = "1.3"
-clap = "2.33"
-rust-gpu-tools = "0.6.0"
+rust-gpu-tools = { version = "0.6.0", default-features = false }
